Revert "Main menu"
