+++ index.html
@@ -10,20 +10,15 @@
             padding: 0;
             overflow: hidden;
             background-color: #000;
-            touch-action: none;
-            user-select: none;
         }
         canvas {
             display: block;
             cursor: none;
-            width: 100%;
-            height: 100%;
-            touch-action: none;
         }
     </style>
 </head>
 <body>
     <canvas id="gameCanvas"></canvas>
-    <script type="module" src="src/main.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
